move useful stuff to utils
